refactor(cat): split entity handlers into create/manage modules
entities.ts hit 515 lines (500-line service limit) after create_circle. Split
along the natural seam — creating new things vs changing existing ones — with
entities.ts kept as the barrel so the handler map is unchanged.
